Add pin height functionality + rework
Calculate measurements based on bodybuilder ratio and height.
Also reworked pin body part and created a template for the form.
Form template is used for the height page. In addition, they both
lead to the same result page but send a different session variable.
Session variable is to tell the page which page sent it and act
accordingly.